requirements.txt
# Core dependencies
crewai>=0.28.0
openai>=1.0.0
python-dotenv>=1.0.0

# Web API
flask>=2.3.0
flask-cors>=4.0.0

# Optional but recommended
serpapi>=0.1.5
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Note: Python 3.8+ is required for CrewAI

